Android: Fix null ptr crash
Happens when proxy is never bound.

BUG=278396
R=qinmin@chromium.org

Review URL: https://codereview.chromium.org/22887048

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@219326 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6c77f3ceef9179801f091c112a22d57d2cf863f9

+++ content/renderer/media/android/stream_texture_factory_android_synchronous_impl.cc
@@ -64,7 +64,7 @@ StreamTextureProxyImpl::~StreamTextureProxyImpl() {}
 
 void StreamTextureProxyImpl::Release() {
   SetClient(NULL);
-  if (!loop_->BelongsToCurrentThread())
+  if (loop_.get() && !loop_->BelongsToCurrentThread())
     loop_->DeleteSoon(FROM_HERE, this);
   else
     delete this;
